New art, animations, scripts, updates on scripts
New art:
Slime sprites including idle and moving sprites.

Animation folder

Slime animation: idle and moving

New scripts:
Bullet Spawn Manager: Keeps a register of all the array numbers for each bullet on screen that are being shared from the object pool.

Bullet Spawn: This is now gonna be in charge of spawning the bullets and accessing their stored rigidbodies along side the Bullet Spawn Manager.

Enemy Attack: This abstract class will be the base class for enemy attacking logic.

Enemy Melee Attack: New script in charge of melee attacks, could be further divided in the future into a specific collider instead of the default.

Enemy Melee Damage was deleted and replaced.

Enemy Shooting: Will use the Enemy Attack logic and alter it to fit range attacks.

Face Player (ON-WORK): This will be added as a component to make the object face the player. Was done this way to make it so that the function gets called only when the player is inside the attack range of long range enemies.

New Prefab:
Water Enemy (ON-WORK): Will be the new long range attack enemy.

New Scriptable Objects:
Created a folder to store all scriptable objects and their old divisions.
Enemy Shooting Stats created, still not filled and used.

Update:
Default enemy is now slime.

Player prefab was updated to contain the new scripts.

Slime prefab was updated to contain the new scripts.

Reorganised the folders for enemies.

Enemy health no longer changes to a specific colour.

Enemy Movement now also works with the enemy moving animation to flip depending on which side the player currently is when chasing.

Object pooling starts now earlier than default. It can now return the pool array number to access a specific pool. It is possible to publicly access the pool size of all pools and it is possible to publicly get the RigidBody2D of prefabs.

Thrown Star script is now Player Bullet and can be used in any bullet used by the player.

Player Shooting was updated to make use of Spawn Bullet script.

Spawner was moved to enemy folder.